relocating RemoveMethodInvocationsVisitor / Test to appropriate dir after porting them over from rewrite-spring with preserved history
